Consolidate KDTreeValues
